@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "modos_schema"
-version = "0.3.1"
+version = "0.3.2"
 description = "Metadata schema for the Multi-Omics Digital Object System"
 authors = ["Cyril Matthey-Doret"]
 license = "Apache-2.0"
@@ -8,11 +8,11 @@ readme = "README.md"
 include = ["README.md", "src/modos_schema/schema", "project"]
 
 [tool.poetry.dependencies]
-python = "^3.9"
-linkml-runtime = "^1.6"
+python = "^3.10"
+linkml-runtime = "^1.9.1"
 
 [tool.poetry.group.dev.dependencies]
-linkml = "^1.8.4"
+linkml = "^1.9.1"
 
 [tool.poetry-dynamic-versioning]
 enable = true
